16/9/12 - Intercept Money mission =======
+ Added a new mission: InterceptMoney
++ Three transfers are set up: anonymous donation into account A, transfer from A to B, withdrawal from account B
++ The player has to siphon off part of the transfer during the window between the first and third transfers, sending it to account C
++ As a result, there is now a scheduler event that can move money around, which I'm sure will prove useful in the future

+ Upgraded SpamGenerator to send some bank scams as well

+ Installed Adaptive Security
++ Monitors, Proxies and Firewalls will creep up in level in response to hacks
++ Uplink automatically produces new bypass and disable versions
++ Note: The cost calculator isn't quite working yet, consider it on the TODO list

+ Added the Encrypter since it rounds out the set of tools added in SoftwareMod

+ Reworked the software categories for SoftwareSales a bit

+ Tidied up SoftwareMod to remove lots and lots of float/int conversion warnings at compile time

+++ uplink/src/app/opengl.cpp
@@ -481,7 +481,7 @@ void keyboard(unsigned char key, int x, int y)
 			app->Close ();
 		} else {
 			bool inmemory = false;
-			int version = 0;
+			float version = 0.0f;
 			DataBank *db = &(game->GetWorld ()->GetPlayer ()->gateway.databank);
 			int numfiles = db->GetDataSize ();
 			for ( int i = 0; i < numfiles; ++i ) {
@@ -598,7 +598,7 @@ void specialkeyboard (int key, int x, int y)
 					app->Close ();
 				} else {
 					bool inmemory = false;
-					int version = 0;
+					float version = 0.0f;
 					DataBank *db = &(game->GetWorld ()->GetPlayer ()->gateway.databank);
 					int numfiles = db->GetDataSize ();
 					for ( int i = 0; i < numfiles; ++i ) {
