+++ uplink/src/interface/taskmanager/decompressor.cpp
@@ -121,31 +121,31 @@ void Decompressor::BorderDraw ( Button *button, bool highlighted, bool clicked )
 
 	glBegin ( GL_QUADS );
 
-		if      ( clicked )		glColor4f ( 0.5, 0.5, 0.6, ALPHA );
-		else if ( highlighted ) glColor4f ( 0.2, 0.2, 0.5, ALPHA );
-		else					glColor4f ( 0.2, 0.2, 0.4, ALPHA );
+		if      ( clicked )		glColor4f ( 0.5f, 0.5f, 0.6f, ALPHA );
+		else if ( highlighted ) glColor4f ( 0.2f, 0.2f, 0.5f, ALPHA );
+		else					glColor4f ( 0.2f, 0.2f, 0.4f, ALPHA );
 		glVertex2i ( button->x, button->y );
 
-		if		( clicked )		glColor4f ( 0.7, 0.7, 0.6, ALPHA );
-		else if ( highlighted ) glColor4f ( 0.5, 0.5, 0.6, ALPHA );
-		else					glColor4f ( 0.3, 0.3, 0.5, ALPHA );
+		if		( clicked )		glColor4f ( 0.7f, 0.7f, 0.6f, ALPHA );
+		else if ( highlighted ) glColor4f ( 0.5f, 0.5f, 0.6f, ALPHA );
+		else					glColor4f ( 0.3f, 0.3f, 0.5f, ALPHA );
 		glVertex2i ( button->x + button->width, button->y );
 
-		if		( clicked )		glColor4f ( 0.5, 0.5, 0.6, ALPHA );
-		else if ( highlighted ) glColor4f ( 0.2, 0.2, 0.5, ALPHA );
-		else					glColor4f ( 0.2, 0.2, 0.4, ALPHA );
+		if		( clicked )		glColor4f ( 0.5f, 0.5f, 0.6f, ALPHA );
+		else if ( highlighted ) glColor4f ( 0.2f, 0.2f, 0.5f, ALPHA );
+		else					glColor4f ( 0.2f, 0.2f, 0.4f, ALPHA );
 		glVertex2i ( button->x + button->width, button->y + button->height );
 
-		if		( clicked )		glColor4f ( 0.7, 0.7, 0.6, ALPHA );
-		else if ( highlighted ) glColor4f ( 0.5, 0.5, 0.6, ALPHA );
-		else					glColor4f ( 0.3, 0.3, 0.5, ALPHA );
+		if		( clicked )		glColor4f ( 0.7f, 0.7f, 0.6f, ALPHA );
+		else if ( highlighted ) glColor4f ( 0.5f, 0.5f, 0.6f, ALPHA );
+		else					glColor4f ( 0.3f, 0.3f, 0.5f, ALPHA );
 		glVertex2i ( button->x, button->y + button->height );
 
 	glEnd ();
 
 	if ( highlighted || clicked ) {
 
-		glColor4f ( 0.3, 0.3, 0.7, 1.0 );
+		glColor4f ( 0.3f, 0.3f, 0.7f, 1.0f );
 		border_draw ( button );
 
 	}
@@ -157,21 +157,21 @@ void Decompressor::ProgressDraw ( Button *button, bool highlighted, bool clicked
 
 	UplinkAssert ( button );
 
-	float scale = (float) button->width / 100.0;
+	float scale = (float) button->width / 100.0f;
 	if ( highlighted ) scale *= 2;
 
 	glBegin ( GL_QUADS );
 
-		glColor4f ( 0.0, 1.5 - scale, scale, 0.6 );
+		glColor4f ( 0.0f, 1.5f - scale, scale, 0.6f );
 		glVertex3i ( button->x, button->y, 0 );
 
-		glColor4f ( 0.0, 1.5 - scale, scale, 0.6 );
+		glColor4f ( 0.0f, 1.5f - scale, scale, 0.6f );
 		glVertex3i ( button->x + button->width, button->y, 0 );
 
-		glColor4f ( 0.0, 1.5 - scale, scale, 0.6 );
+		glColor4f ( 0.0f, 1.5f - scale, scale, 0.6f );
 		glVertex3i ( button->x + button->width, button->y + button->height, 0 );
 
-		glColor4f ( 0.0, 1.5 - scale, scale, 0.6 );
+		glColor4f ( 0.0f, 1.5f - scale, scale, 0.6f );
 		glVertex3i ( button->x, button->y + button->height, 0 );
 
 	glEnd ();
@@ -295,7 +295,7 @@ void Decompressor::Tick ( int n )
 				++progress;		
 
 				UplinkAssert ( EclGetButton ( sprogress ) );
-				EclGetButton ( sprogress )->width = 120 * ( (float) progress / (float) numticksrequired );
+				EclGetButton ( sprogress )->width = (int) (120 * ( (float) progress / (float) numticksrequired ));
 				EclDirtyButton ( sprogress );
 
 				if ( progress >= numticksrequired ) {
