+++ uplink/src/world/generator/langenerator.cpp
@@ -120,9 +120,9 @@ Computer  *LanGenerator::LoadLAN ( char *filename )
 
 	Company *company = game->GetWorld ()->GetCompany( companyName );
 	if ( !company ) {
-		int size      = NumberGenerator::RandomNormalNumber ( COMPANYSIZE_AVERAGE,      COMPANYSIZE_RANGE );
-		int growth    = NumberGenerator::RandomNormalNumber ( COMPANYGROWTH_AVERAGE,    COMPANYGROWTH_RANGE );
-		int alignment = NumberGenerator::RandomNormalNumber ( COMPANYALIGNMENT_AVERAGE, COMPANYALIGNMENT_RANGE );
+		int size      = (int) NumberGenerator::RandomNormalNumber ( COMPANYSIZE_AVERAGE,      COMPANYSIZE_RANGE );
+		int growth    = (int) NumberGenerator::RandomNormalNumber ( COMPANYGROWTH_AVERAGE,    COMPANYGROWTH_RANGE );
+		int alignment = (int) NumberGenerator::RandomNormalNumber ( COMPANYALIGNMENT_AVERAGE, COMPANYALIGNMENT_RANGE );
     		company = WorldGenerator::GenerateCompany ( companyName, size, COMPANYTYPE_COMMERCIAL, growth, alignment );
     
 		Person *ceo = WorldGenerator::GetRandomPerson ();
